refactor: replace AdminDashboard wrapper with full component
